requirements/tests.txt
django>=1.5.0
django_nose>=1.1.0
coverage>=3.6
glob2
